+++ Changelog.md
@@ -12,6 +12,7 @@
 - Clarified Smart Consumption tooltips
 - Clarified Can Consume Itself to Smart Consumption
 - Added Quick Stack to Smart Pickup
+- Added Consumption to Return to Previous Slot
 - Fixed Quick with Frame Skip Off
 - Fixed Extra Item Right Click conflicting with Terraria Overhaul